fix: build error

+++ source/MaaFramework/Task/Component/Recognizer.cpp
@@ -325,7 +325,7 @@ RecoResult Recognizer::and_(const std::shared_ptr<MAA_RES_NS::Recognition::AndPa
         return result;
     }
 
-    if (sub_results.size() <= param->box_index) {
+    if (static_cast<int>(sub_results.size()) <= param->box_index) {
         LogError << "all hit, but box_index is out of range" << VAR(name) << VAR(sub_results.size()) << VAR(param->box_index);
         return {};
     }
